+++ crates/amaru-mithril/src/lib.rs
@@ -23,10 +23,10 @@ pub use archive::{
     ArchiveMetadata, BLOCKS_PER_ARCHIVE, archive_name_for_blocks, latest_archive, list_existing_archives,
     package_blocks, parse_archive_metadata, resume_point_for_archives, sorted_archives,
 };
-pub use download::download_from_mithril;
+pub use download::{download_from_mithril, download_from_mithril_for_resume_point};
 pub use immutable::{
     ImmutableBlock, ImmutableBlocksIter, chunk_for_slot, first_missing_immutable_chunk, from_chunk_for_resume_point,
-    get_latest_chunk, iter_immutable_blocks, read_stable_blocks_after_point,
+    get_latest_chunk, iter_immutable_blocks, read_blocks_after_point, read_stable_blocks_after_point,
 };
 
 #[derive(Debug)]